+++ CHANGELOG.md
@@ -6,7 +6,10 @@ TableTransform is still in pre-release state. This means that its APIs and behav
 
 ## [Unreleased]
 
-* Added format capability for columns when published as Excel 
+* Added format capability for columns when published as Excel
+* CodeClimate and test coverage added
+* Ruby >= 2.1 verified with TravisCI
+* Improved documentation
 
 ## [0.2.0][] (2016-02-16)