+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 TableTransform is still in pre-release state. This means that its APIs and behavior are subject to breaking changes without deprecation notices. Until 1.0, version numbers will follow a [Semver][]-ish `0.y.z` format, where `y` is incremented when new features or breaking changes are introduced, and `z` is incremented for lesser changes or bug fixes.
 
 ## [Unreleased]
+
+## [0.4.0][] (2016-03-05)
 * Supports formulas in columns
 * Helper functions to create formulas
 * Column width estimates format size in calculation 
@@ -24,5 +26,6 @@ TableTransform is still in pre-release state. This means that its APIs and behav
 
 [Semver]: http://semver.org
 [Unreleased]: https://github.com/jonas-lantto/table_transform/compare/v0.3.0...HEAD
+[0.4.0]: https://github.com/jonas-lantto/table_transform/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/jonas-lantto/table_transform/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/jonas-lantto/table_transform/compare/v0.1.0...v0.2.0